+++ modules/apps/site/site-admin-web/src/com/liferay/site/admin/web/servlet/taglib/ui/SiteSitemapFormNavigatorEntry.java
@@ -12,13 +12,11 @@
  * details.
  */
 
-package com.liferay.site.admin.web.taglib.ui;
+package com.liferay.site.admin.web.servlet.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
-import com.liferay.portal.kernel.util.PrefsPropsUtil;
-import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.User;
 
@@ -33,35 +31,26 @@ import org.osgi.service.component.annotations.Component;
 	property = {"service.ranking:Integer=20"},
 	service = FormNavigatorEntry.class
 )
-public class SiteContentSharingFormNavigatorEntry
-	extends BaseSiteFormNavigatorEntry {
+public class SiteSitemapFormNavigatorEntry extends BaseSiteFormNavigatorEntry {
 
 	@Override
 	public String getCategoryKey() {
-		return FormNavigatorConstants.CATEGORY_KEY_SITES_ADVANCED;
+		return FormNavigatorConstants.CATEGORY_KEY_SITES_SEO;
 	}
 
 	@Override
 	public String getKey() {
-		return "content-sharing";
+		return "sitemap";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "content-sharing");
+		return LanguageUtil.get(locale, "sitemap");
 	}
 
 	@Override
 	public boolean isVisible(User user, Group group) {
-		if (group == null) {
-			return false;
-		}
-
-		int contentSharingWithChildrenEnabled = PrefsPropsUtil.getInteger(
-			group.getCompanyId(),
-			PropsKeys.SITES_CONTENT_SHARING_WITH_CHILDREN_ENABLED);
-
-		if (contentSharingWithChildrenEnabled == 0) {
+		if ((group == null) || group.isCompany()) {
 			return false;
 		}
 
@@ -70,7 +59,7 @@ public class SiteContentSharingFormNavigatorEntry
 
 	@Override
 	protected String getJspPath() {
-		return "/site/content_sharing.jsp";
+		return "/site/sitemap.jsp";
 	}
 
 }
